Conflicts:
	pom.xml
	xchange-coinfloor/pom.xml
	xchange-coinfloor/src/main/java/com/xeiam/xchange/coinfloor/CoinfloorExchange.java
	xchange-coinfloor/src/main/java/com/xeiam/xchange/coinfloor/CoinfloorUtils.java
	xchange-coinfloor/src/main/java/com/xeiam/xchange/coinfloor/streaming/CoinfloorEventListener.java
	xchange-coinfloor/src/main/java/com/xeiam/xchange/coinfloor/streaming/CoinfloorStreamingExchangeService.java
	xchange-examples/src/main/java/com/xeiam/xchange/examples/coinfloor/CoinfloorDemo.java
	xchange-examples/src/main/java/com/xeiam/xchange/examples/coinfloor/CoinfloorExampleUtils.java

+++ xchange-cexio/src/main/java/com/xeiam/xchange/cexio/dto/trade/CexIOOrder.java
@@ -1,24 +1,3 @@
-/**
- * Copyright (C) 2012 - 2014 Xeiam LLC http://xeiam.com
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of
- * this software and associated documentation files (the "Software"), to deal in
- * the Software without restriction, including without limitation the rights to
- * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies
- * of the Software, and to permit persons to whom the Software is furnished to do
- * so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
 package com.xeiam.xchange.cexio.dto.trade;
 
 import java.math.BigDecimal;
@@ -31,7 +10,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
  */
 public class CexIOOrder {
 
-  private final int id;
+  private final long id;
   private final long time;
   private final Type type;
   private final BigDecimal price;
@@ -53,7 +32,7 @@ public class CexIOOrder {
    * @param amount
    * @param pending
    */
-  public CexIOOrder(@JsonProperty("id") int id, @JsonProperty("time") long time, @JsonProperty("type") Type type, @JsonProperty("price") BigDecimal price, @JsonProperty("amount") BigDecimal amount,
+  public CexIOOrder(@JsonProperty("id") long id, @JsonProperty("time") long time, @JsonProperty("type") Type type, @JsonProperty("price") BigDecimal price, @JsonProperty("amount") BigDecimal amount,
       @JsonProperty("pending") BigDecimal pending, @JsonProperty("error") String errorMessage) {
 
     this.id = id;
@@ -65,7 +44,7 @@ public class CexIOOrder {
     this.errorMessage = errorMessage;
   }
 
-  public int getId() {
+  public long getId() {
 
     return id;
   }
